+++ rust/main/agents/relayer/src/relayer.rs
@@ -70,10 +70,10 @@ const CURSOR_BUILDING_ERROR: &str = "Error building cursor for origin";
 const CURSOR_INSTANTIATION_ATTEMPTS: usize = 10;
 const ADVANCED_LOG_META: bool = false;
 
-#[derive(Debug, Hash, PartialEq, Eq, Copy, Clone)]
+#[derive(Debug, Hash, PartialEq, Eq, Clone)]
 struct ContextKey {
-    origin: u32,
-    destination: u32,
+    origin: HyperlaneDomain,
+    destination: HyperlaneDomain,
 }
 
 /// A relayer agent
@@ -154,10 +154,11 @@ impl BaseAgent for Relayer {
         Self::reset_critical_errors(&settings, &chain_metrics);
 
         let start = Instant::now();
-        let mut start_entity_init = Instant::now();
 
         let core = settings.build_hyperlane_core(core_metrics.clone());
-        let db = DB::from_path(&settings.db)?;
+
+        let mut start_entity_init = Instant::now();
+
         let cache_name = "relayer_cache";
         let inner_cache = if settings.allow_contract_call_caching {
             Some(MeteredCache::new(
@@ -171,6 +172,10 @@ impl BaseAgent for Relayer {
             None
         };
         let cache = OptionalCache::new(inner_cache);
+        debug!(elapsed = ?start_entity_init.elapsed(), event = "initialized cache", "Relayer startup duration measurement");
+
+        start_entity_init = Instant::now();
+        let db = DB::from_path(&settings.db)?;
         let dbs = settings
             .origin_chains
             .iter()
@@ -217,15 +222,12 @@ impl BaseAgent for Relayer {
         .await;
         debug!(elapsed = ?start_entity_init.elapsed(), event = "initialized dipatchers", "Relayer startup duration measurement");
 
-        let contract_sync_metrics = Arc::new(ContractSyncMetrics::new(&core_metrics));
-
         start_entity_init = Instant::now();
-
+        let contract_sync_metrics = Arc::new(ContractSyncMetrics::new(&core_metrics));
         let stores: HashMap<_, _> = dbs
             .iter()
             .map(|(d, db)| (d.clone(), Arc::new(db.clone())))
             .collect();
-
         let message_syncs = Self::build_contract_syncs(
             &settings,
             &core_metrics,
@@ -235,7 +237,6 @@ impl BaseAgent for Relayer {
             "message",
         )
         .await;
-
         debug!(elapsed = ?start_entity_init.elapsed(), event = "initialized message syncs", "Relayer startup duration measurement");
 
         start_entity_init = Instant::now();
@@ -307,10 +308,10 @@ impl BaseAgent for Relayer {
             .iter()
             .filter_map(|domain| match dbs.get(domain) {
                 Some(db) => {
-                    let gas_payment_enforcer = Arc::new(GasPaymentEnforcer::new(
+                    let gas_payment_enforcer = Arc::new(RwLock::new(GasPaymentEnforcer::new(
                         settings.gas_payment_enforcement.clone(),
                         db.clone(),
-                    ));
+                    )));
                     Some((domain.clone(), gas_payment_enforcer))
                 }
                 None => {
@@ -321,41 +322,48 @@ impl BaseAgent for Relayer {
             .collect();
         debug!(elapsed = ?start_entity_init.elapsed(), event = "initialized gas payment enforcers", "Relayer startup duration measurement");
 
-        let mut msg_ctxs = HashMap::new();
-
-        start_entity_init = Instant::now();
-
         // only iterate through destination chains that were successfully instantiated
-        let ccip_signer_futures = mailboxes
-            .keys()
-            .map(|destination| {
-                let destination_chain_setup =
-                    core.settings.chain_setup(destination).unwrap().clone();
-                let signer = destination_chain_setup.signer.clone();
-                async move {
-                    if !matches!(destination.domain_protocol(), HyperlaneDomainProtocol::Ethereum) {
-                        return (destination, None);
-                    }
-                    let signer = if let Some(builder) = signer {
-                        match builder.build::<hyperlane_ethereum::Signers>().await {
-                            Ok(signer) => Some(signer),
-                            Err(err) => {
-                                warn!(error = ?err, "Failed to build Ethereum signer for CCIP-read ISM. ");
-                                None
-                            }
-                        }
-                    } else {
-                        None
-                    };
-                    (destination, signer)
+        start_entity_init = Instant::now();
+        let mut ccip_signer_futures: Vec<_> = Vec::with_capacity(mailboxes.len());
+        for destination in mailboxes.keys() {
+            let destination_chain_setup = match core.settings.chain_setup(destination) {
+                Ok(setup) => setup.clone(),
+                Err(err) => {
+                    tracing::error!(?destination, ?err, "Destination chain setup failed");
+                    continue;
                 }
-            })
-            .collect::<Vec<_>>();
+            };
+            let signer = destination_chain_setup.signer.clone();
+            let future = async move {
+                if !matches!(
+                    destination.domain_protocol(),
+                    HyperlaneDomainProtocol::Ethereum
+                ) {
+                    return (destination, None);
+                }
+                let signer = if let Some(builder) = signer {
+                    match builder.build::<hyperlane_ethereum::Signers>().await {
+                        Ok(signer) => Some(signer),
+                        Err(err) => {
+                            warn!(error = ?err, "Failed to build Ethereum signer for CCIP-read ISM. ");
+                            None
+                        }
+                    }
+                } else {
+                    None
+                };
+                (destination, signer)
+            };
+            ccip_signer_futures.push(future);
+        }
         let ccip_signers = join_all(ccip_signer_futures)
             .await
             .into_iter()
             .collect::<HashMap<_, _>>();
+        debug!(elapsed = ?start_entity_init.elapsed(), event = "initialized ccip signers", "Relayer startup duration measurement");
 
+        start_entity_init = Instant::now();
+        let mut msg_ctxs = HashMap::new();
         let mut destination_chains = HashMap::new();
         for (destination, dest_mailbox) in mailboxes.iter() {
             let destination_chain_setup = match core.settings.chain_setup(destination) {
@@ -416,8 +424,8 @@ impl BaseAgent for Relayer {
 
                 msg_ctxs.insert(
                     ContextKey {
-                        origin: origin.id(),
-                        destination: destination.id(),
+                        origin: origin.clone(),
+                        destination: destination.clone(),
                     },
                     Arc::new(MessageContext {
                         destination_mailbox: dest_mailbox.clone(),
@@ -512,21 +520,35 @@ impl BaseAgent for Relayer {
             };
 
             // Default to submitting one message at a time if there is no batch config
-            let match_batch_size = self.core.settings.chains[dest_domain.name()]
-                .connection
-                .operation_submission_config()
-                .map(|c| c.max_batch_size)
+            let max_batch_size = self
+                .core
+                .settings
+                .chains
+                .get(dest_domain.name())
+                .and_then(|chain| {
+                    chain
+                        .connection
+                        .operation_submission_config()
+                        .map(|c| c.max_batch_size)
+                })
                 .unwrap_or(1);
-            let max_submit_queue_len = self.core.settings.chains[dest_domain.name()]
-                .connection
-                .operation_submission_config()
-                .and_then(|c| c.max_submit_queue_length);
+            let max_submit_queue_len =
+                self.core
+                    .settings
+                    .chains
+                    .get(dest_domain.name())
+                    .and_then(|chain| {
+                        chain
+                            .connection
+                            .operation_submission_config()
+                            .and_then(|c| c.max_submit_queue_length)
+                    });
             let serial_submitter = SerialSubmitter::new(
                 dest_domain.clone(),
                 receive_channel,
                 &sender,
                 SerialSubmitterMetrics::new(&self.core.metrics, dest_domain),
-                match_batch_size,
+                max_batch_size,
                 max_submit_queue_len,
                 task_monitor.clone(),
                 payload_dispatcher_entrypoint,
@@ -544,7 +566,7 @@ impl BaseAgent for Relayer {
                 tasks.push(dispatcher.spawn().await);
             }
 
-            let metrics_updater = ChainSpecificMetricsUpdater::new(
+            let metrics_updater = match ChainSpecificMetricsUpdater::new(
                 dest_conf,
                 self.core_metrics.clone(),
                 self.agent_metrics.clone(),
@@ -552,9 +574,18 @@ impl BaseAgent for Relayer {
                 Self::AGENT_NAME.to_string(),
             )
             .await
-            .unwrap_or_else(|_| {
-                panic!("Error creating metrics updater for destination {dest_domain}")
-            });
+            {
+                Ok(task) => task,
+                Err(err) => {
+                    Self::record_critical_error(
+                        dest_domain,
+                        &self.chain_metrics,
+                        &err,
+                        "Failed to build metrics updater",
+                    );
+                    continue;
+                }
+            };
             tasks.push(metrics_updater.spawn());
         }
         debug!(elapsed = ?start_entity_init.elapsed(), event = "started submitters", "Relayer startup duration measurement");
@@ -565,32 +596,98 @@ impl BaseAgent for Relayer {
                 .message_syncs
                 .get(origin)
                 .and_then(|sync| sync.get_broadcaster());
-            tasks.push(self.run_message_sync(origin, task_monitor.clone()).await);
+
+            let message_sync = match self.run_message_sync(origin, task_monitor.clone()).await {
+                Ok(task) => task,
+                Err(err) => {
+                    Self::record_critical_error(
+                        origin,
+                        &self.chain_metrics,
+                        &err,
+                        "Failed to run message sync",
+                    );
+                    continue;
+                }
+            };
+            tasks.push(message_sync);
+
             if let Some(interchain_gas_payment_syncs) = &self.interchain_gas_payment_syncs {
-                tasks.push(
-                    self.run_interchain_gas_payment_sync(
+                let interchain_gas_payment_sync = match self
+                    .run_interchain_gas_payment_sync(
                         origin,
                         interchain_gas_payment_syncs,
                         BroadcastMpscSender::map_get_receiver(maybe_broadcaster.as_ref()).await,
                         task_monitor.clone(),
                     )
-                    .await,
-                );
+                    .await
+                {
+                    Ok(task) => task,
+                    Err(err) => {
+                        Self::record_critical_error(
+                            origin,
+                            &self.chain_metrics,
+                            &err,
+                            "Failed to run interchain gas payment sync",
+                        );
+                        continue;
+                    }
+                };
+                tasks.push(interchain_gas_payment_sync);
             }
-            tasks.push(
-                self.run_merkle_tree_hook_sync(
+
+            let merkle_tree_hook_sync = match self
+                .run_merkle_tree_hook_sync(
                     origin,
                     BroadcastMpscSender::map_get_receiver(maybe_broadcaster.as_ref()).await,
                     task_monitor.clone(),
                 )
-                .await,
-            );
-            tasks.push(self.run_message_processor(
+                .await
+            {
+                Ok(task) => task,
+                Err(err) => {
+                    Self::record_critical_error(
+                        origin,
+                        &self.chain_metrics,
+                        &err,
+                        "Failed to run merkle tree hook sync",
+                    );
+                    continue;
+                }
+            };
+            tasks.push(merkle_tree_hook_sync);
+
+            let message_processor = match self.run_message_processor(
                 origin,
                 send_channels.clone(),
                 task_monitor.clone(),
-            ));
-            tasks.push(self.run_merkle_tree_processor(origin, task_monitor.clone()));
+            ) {
+                Ok(task) => task,
+                Err(err) => {
+                    Self::record_critical_error(
+                        origin,
+                        &self.chain_metrics,
+                        &err,
+                        "Failed to run message processor",
+                    );
+                    continue;
+                }
+            };
+            tasks.push(message_processor);
+
+            let merkle_tree_processor =
+                match self.run_merkle_tree_processor(origin, task_monitor.clone()) {
+                    Ok(task) => task,
+                    Err(err) => {
+                        Self::record_critical_error(
+                            origin,
+                            &self.chain_metrics,
+                            &err,
+                            "Failed to run merkle tree processor",
+                        );
+                        continue;
+                    }
+                };
+            tasks.push(merkle_tree_processor);
         }
         debug!(elapsed = ?start_entity_init.elapsed(), event = "started message, IGP, merkle tree hook syncs, and message and merkle tree processors", "Relayer startup duration measurement");
 
@@ -600,10 +697,17 @@ impl BaseAgent for Relayer {
         // create a db mapping for server handlers
         let dbs: HashMap<u32, HyperlaneRocksDB> =
             self.dbs.iter().map(|(k, v)| (k.id(), v.clone())).collect();
+
+        let gas_enforcers: HashMap<_, _> = self
+            .msg_ctxs
+            .iter()
+            .map(|(key, ctx)| (key.origin.clone(), ctx.origin_gas_payment_enforcer.clone()))
+            .collect();
         let relayer_router = relayer_server::Server::new(self.destination_chains.len())
             .with_op_retry(sender.clone())
             .with_message_queue(prep_queues)
             .with_dbs(dbs)
+            .with_gas_enforcers(gas_enforcers)
             .router();
 
         let server = self
@@ -667,14 +771,18 @@ impl Relayer {
         &self,
         origin: &HyperlaneDomain,
         task_monitor: TaskMonitor,
-    ) -> JoinHandle<()> {
+    ) -> eyre::Result<JoinHandle<()>> {
         let origin = origin.clone();
-        let contract_sync = self.message_syncs.get(&origin).unwrap().clone();
+        let contract_sync = self
+            .message_syncs
+            .get(&origin)
+            .cloned()
+            .ok_or_else(|| eyre::eyre!("No message sync found"))?;
         let index_settings = self.as_ref().settings.chains[origin.name()].index_settings();
         let chain_metrics = self.chain_metrics.clone();
 
         let name = Self::contract_sync_task_name("message::", origin.name());
-        tokio::task::Builder::new()
+        Ok(tokio::task::Builder::new()
             .name(&name)
             .spawn(TaskMonitor::instrument(
                 &task_monitor,
@@ -684,7 +792,7 @@ impl Relayer {
                 }
                 .instrument(info_span!("MessageSync")),
             ))
-            .expect("spawning tokio task from Builder is infallible")
+            .expect("spawning tokio task from Builder is infallible"))
     }
 
     async fn message_sync_task(
@@ -716,14 +824,23 @@ impl Relayer {
         >,
         tx_id_receiver: Option<MpscReceiver<H512>>,
         task_monitor: TaskMonitor,
-    ) -> JoinHandle<()> {
+    ) -> eyre::Result<JoinHandle<()>> {
         let origin = origin.clone();
-        let index_settings = self.as_ref().settings.chains[origin.name()].index_settings();
-        let contract_sync = interchain_gas_payment_syncs.get(&origin).unwrap().clone();
+        let index_settings = self
+            .as_ref()
+            .settings
+            .chains
+            .get(origin.name())
+            .map(|settings| settings.index_settings())
+            .ok_or_else(|| eyre::eyre!("Error finding chain index settings"))?;
+        let contract_sync = interchain_gas_payment_syncs
+            .get(&origin)
+            .cloned()
+            .ok_or_else(|| eyre::eyre!("No interchain gas payment sync found"))?;
         let chain_metrics = self.chain_metrics.clone();
 
         let name = Self::contract_sync_task_name("gas_payment::", origin.name());
-        tokio::task::Builder::new()
+        Ok(tokio::task::Builder::new()
             .name(&name)
             .spawn(TaskMonitor::instrument(
                 &task_monitor,
@@ -739,7 +856,7 @@ impl Relayer {
                 }
                 .instrument(info_span!("IgpSync")),
             ))
-            .expect("spawning tokio task from Builder is infallible")
+            .expect("spawning tokio task from Builder is infallible"))
     }
 
     async fn interchain_gas_payments_sync_task(
@@ -772,15 +889,26 @@ impl Relayer {
         origin: &HyperlaneDomain,
         tx_id_receiver: Option<MpscReceiver<H512>>,
         task_monitor: TaskMonitor,
-    ) -> JoinHandle<()> {
+    ) -> eyre::Result<JoinHandle<()>> {
         let origin = origin.clone();
-        let index_settings = self.as_ref().settings.chains[origin.name()].index.clone();
-        let contract_sync = self.merkle_tree_hook_syncs.get(&origin).unwrap().clone();
         let chain_metrics = self.chain_metrics.clone();
+
+        let index_settings = self
+            .as_ref()
+            .settings
+            .chains
+            .get(origin.name())
+            .map(|settings| settings.index_settings())
+            .ok_or_else(|| eyre::eyre!("Error finding chain index settings"))?;
+        let contract_sync = self
+            .merkle_tree_hook_syncs
+            .get(&origin)
+            .cloned()
+            .ok_or_else(|| eyre::eyre!("No merkle tree hook sync found"))?;
 
         let origin_name = origin.name().to_string();
         let name = Self::contract_sync_task_name("merkle_tree::", &origin_name);
-        tokio::task::Builder::new()
+        Ok(tokio::task::Builder::new()
             .name(&name)
             .spawn(TaskMonitor::instrument(
                 &task_monitor,
@@ -796,7 +924,7 @@ impl Relayer {
                 }
                 .instrument(info_span!("MerkleTreeHookSync")),
             ))
-            .expect("spawning tokio task from Builder is infallible")
+            .expect("spawning tokio task from Builder is infallible"))
     }
 
     async fn merkle_tree_hook_sync_task(
@@ -833,7 +961,7 @@ impl Relayer {
         origin: &HyperlaneDomain,
         send_channels: HashMap<u32, UnboundedSender<QueueOperation>>,
         task_monitor: TaskMonitor,
-    ) -> JoinHandle<()> {
+    ) -> eyre::Result<JoinHandle<()>> {
         let metrics = MessageProcessorMetrics::new(
             &self.core.metrics,
             origin,
@@ -844,8 +972,8 @@ impl Relayer {
             .keys()
             .filter_map(|destination| {
                 let key = ContextKey {
-                    origin: origin.id(),
-                    destination: destination.id(),
+                    origin: origin.clone(),
+                    destination: destination.clone(),
                 };
                 let context = self
                     .msg_ctxs
@@ -870,8 +998,14 @@ impl Relayer {
             })
             .collect();
 
+        let db = self
+            .dbs
+            .get(origin)
+            .cloned()
+            .ok_or_else(|| eyre::eyre!("Db not found"))?;
+
         let message_processor = MessageProcessor::new(
-            self.dbs.get(origin).unwrap().clone(),
+            db,
             self.message_whitelist.clone(),
             self.message_blacklist.clone(),
             self.address_blacklist.clone(),
@@ -884,25 +1018,30 @@ impl Relayer {
 
         let span = info_span!("MessageProcessor", origin=%message_processor.domain());
         let processor = Processor::new(Box::new(message_processor), task_monitor.clone());
-
-        processor.spawn(span)
+        Ok(processor.spawn(span))
     }
 
     fn run_merkle_tree_processor(
         &self,
         origin: &HyperlaneDomain,
         task_monitor: TaskMonitor,
-    ) -> JoinHandle<()> {
+    ) -> eyre::Result<JoinHandle<()>> {
         let metrics = MerkleTreeProcessorMetrics::new(&self.core.metrics, origin);
-        let merkle_tree_processor = MerkleTreeProcessor::new(
-            self.dbs.get(origin).unwrap().clone(),
-            metrics,
-            self.prover_syncs[origin].clone(),
-        );
+        let db = self
+            .dbs
+            .get(origin)
+            .cloned()
+            .ok_or_else(|| eyre::eyre!("Db not found"))?;
+        let prover_sync = self
+            .prover_syncs
+            .get(origin)
+            .cloned()
+            .ok_or_else(|| eyre::eyre!("No prover sync found"))?;
 
+        let merkle_tree_processor = MerkleTreeProcessor::new(db, metrics, prover_sync);
         let span = info_span!("MerkleTreeProcessor", origin=%merkle_tree_processor.domain());
         let processor = Processor::new(Box::new(merkle_tree_processor), task_monitor.clone());
-        processor.spawn(span)
+        Ok(processor.spawn(span))
     }
 
     #[allow(clippy::too_many_arguments)]
@@ -976,7 +1115,13 @@ impl Relayer {
         let entrypoint_futures: Vec<_> = settings
             .destination_chains
             .iter()
-            .filter(|chain| SubmitterType::Lander == settings.chains[&chain.to_string()].submitter)
+            .filter(|chain| {
+                settings
+                    .chains
+                    .get(&chain.to_string())
+                    .map(|chain| chain.submitter == SubmitterType::Lander)
+                    .unwrap_or(false)
+            })
             .map(|chain| {
                 (
                     chain.clone(),
@@ -1028,7 +1173,13 @@ impl Relayer {
         let dispatcher_futures: Vec<_> = settings
             .destination_chains
             .iter()
-            .filter(|chain| SubmitterType::Lander == settings.chains[&chain.to_string()].submitter)
+            .filter(|chain| {
+                settings
+                    .chains
+                    .get(&chain.to_string())
+                    .map(|chain| chain.submitter == SubmitterType::Lander)
+                    .unwrap_or(false)
+            })
             .map(|chain| {
                 (
                     chain.clone(),
@@ -1105,7 +1256,7 @@ impl Relayer {
         chain_metrics: &ChainMetrics,
     ) -> HashMap<HyperlaneDomain, Arc<dyn ApplicationOperationVerifier>> {
         settings
-            .build_application_operation_verifiers(settings.origin_chains.iter(), core_metrics)
+            .build_application_operation_verifiers(settings.destination_chains.iter(), core_metrics)
             .await
             .into_iter()
             .filter_map(
@@ -1175,237 +1326,4 @@ impl Relayer {
 }
 
 #[cfg(test)]
-mod test {
-    use std::{
-        collections::{HashMap, HashSet},
-        path::PathBuf,
-        time::Duration,
-    };
-
-    use ethers::utils::hex;
-    use ethers_prometheus::middleware::PrometheusMiddlewareConf;
-    use prometheus::{opts, IntGaugeVec, Registry};
-    use reqwest::Url;
-
-    use hyperlane_base::{
-        settings::{
-            ChainConf, ChainConnectionConf, CoreContractAddresses, IndexSettings, Settings,
-            TracingConfig,
-        },
-        ChainMetrics, CoreMetrics, BLOCK_HEIGHT_HELP, BLOCK_HEIGHT_LABELS, CRITICAL_ERROR_HELP,
-        CRITICAL_ERROR_LABELS,
-    };
-    use hyperlane_core::{
-        config::OpSubmissionConfig, HyperlaneDomain, IndexMode, KnownHyperlaneDomain, ReorgPeriod,
-        H256,
-    };
-    use hyperlane_ethereum as h_eth;
-
-    use crate::settings::{matching_list::MatchingList, RelayerSettings};
-
-    use super::Relayer;
-
-    /// Builds a test RelayerSetting
-    fn generate_test_relayer_settings() -> RelayerSettings {
-        let chains = [(
-            "arbitrum".to_string(),
-            ChainConf {
-                domain: HyperlaneDomain::Known(KnownHyperlaneDomain::Arbitrum),
-                signer: None,
-                submitter: Default::default(),
-                estimated_block_time: Duration::from_secs_f64(1.1),
-                reorg_period: ReorgPeriod::None,
-                addresses: CoreContractAddresses {
-                    mailbox: H256::from_slice(
-                        hex::decode(
-                            "000000000000000000000000598facE78a4302f11E3de0bee1894Da0b2Cb71F8",
-                        )
-                        .unwrap()
-                        .as_slice(),
-                    ),
-                    interchain_gas_paymaster: H256::from_slice(
-                        hex::decode(
-                            "000000000000000000000000c756cFc1b7d0d4646589EDf10eD54b201237F5e8",
-                        )
-                        .unwrap()
-                        .as_slice(),
-                    ),
-                    validator_announce: H256::from_slice(
-                        hex::decode(
-                            "0000000000000000000000001b33611fCc073aB0737011d5512EF673Bff74962",
-                        )
-                        .unwrap()
-                        .as_slice(),
-                    ),
-                    merkle_tree_hook: H256::from_slice(
-                        hex::decode(
-                            "000000000000000000000000AD34A66Bf6dB18E858F6B686557075568c6E031C",
-                        )
-                        .unwrap()
-                        .as_slice(),
-                    ),
-                },
-                connection: ChainConnectionConf::Ethereum(h_eth::ConnectionConf {
-                    rpc_connection: h_eth::RpcConnectionConf::Http {
-                        url: Url::parse("https://sepolia-rollup.arbitrum.io/rpc").unwrap(),
-                    },
-                    transaction_overrides: h_eth::TransactionOverrides {
-                        gas_price: None,
-                        gas_limit: None,
-                        max_fee_per_gas: None,
-                        max_priority_fee_per_gas: None,
-                        ..Default::default()
-                    },
-                    op_submission_config: OpSubmissionConfig {
-                        batch_contract_address: None,
-                        max_batch_size: 1,
-                        ..Default::default()
-                    },
-                }),
-                metrics_conf: PrometheusMiddlewareConf {
-                    contracts: HashMap::new(),
-                    chain: None,
-                },
-                index: IndexSettings {
-                    from: 0,
-                    chunk_size: 1,
-                    mode: IndexMode::Block,
-                },
-                ignore_reorg_reports: false,
-            },
-        )];
-
-        RelayerSettings {
-            base: Settings {
-                chains: chains.into_iter().collect(),
-                metrics_port: 5000,
-                tracing: TracingConfig::default(),
-            },
-            db: PathBuf::new(),
-            origin_chains: [
-                HyperlaneDomain::Known(KnownHyperlaneDomain::Arbitrum),
-                HyperlaneDomain::Known(KnownHyperlaneDomain::Ethereum),
-                HyperlaneDomain::Known(KnownHyperlaneDomain::Optimism),
-            ]
-            .into_iter()
-            .collect(),
-            destination_chains: [
-                HyperlaneDomain::Known(KnownHyperlaneDomain::Arbitrum),
-                HyperlaneDomain::Known(KnownHyperlaneDomain::Ethereum),
-                HyperlaneDomain::Known(KnownHyperlaneDomain::Optimism),
-            ]
-            .into_iter()
-            .collect(),
-            gas_payment_enforcement: Vec::new(),
-            whitelist: MatchingList::default(),
-            blacklist: MatchingList::default(),
-            address_blacklist: Vec::new(),
-            transaction_gas_limit: None,
-            skip_transaction_gas_limit_for: HashSet::new(),
-            allow_local_checkpoint_syncers: true,
-            metric_app_contexts: Vec::new(),
-            allow_contract_call_caching: true,
-            ism_cache_configs: Default::default(),
-            max_retries: 1,
-            tx_id_indexing_enabled: true,
-            igp_indexing_enabled: true,
-        }
-    }
-
-    #[tokio::test]
-    #[tracing_test::traced_test]
-    async fn test_failed_build_mailboxes() {
-        let settings = generate_test_relayer_settings();
-
-        let registry = Registry::new();
-        let core_metrics = CoreMetrics::new("relayer", 4000, registry).unwrap();
-        let chain_metrics = ChainMetrics {
-            block_height: IntGaugeVec::new(
-                opts!("block_height", BLOCK_HEIGHT_HELP),
-                BLOCK_HEIGHT_LABELS,
-            )
-            .unwrap(),
-            gas_price: None,
-            critical_error: IntGaugeVec::new(
-                opts!("critical_error", CRITICAL_ERROR_HELP),
-                CRITICAL_ERROR_LABELS,
-            )
-            .unwrap(),
-        };
-
-        let mailboxes = Relayer::build_mailboxes(&settings, &core_metrics, &chain_metrics).await;
-
-        assert_eq!(mailboxes.len(), 1);
-        assert!(mailboxes.contains_key(&HyperlaneDomain::Known(KnownHyperlaneDomain::Arbitrum)));
-
-        // Arbitrum chain should not have any errors because it's ChainConf exists
-        let metric = chain_metrics
-            .critical_error
-            .get_metric_with_label_values(&["arbitrum"])
-            .unwrap();
-        assert_eq!(metric.get(), 0);
-
-        // Ethereum chain should error because it is missing ChainConf
-        let metric = chain_metrics
-            .critical_error
-            .get_metric_with_label_values(&["ethereum"])
-            .unwrap();
-        assert_eq!(metric.get(), 1);
-
-        // Optimism chain should error because it is missing ChainConf
-        let metric = chain_metrics
-            .critical_error
-            .get_metric_with_label_values(&["optimism"])
-            .unwrap();
-        assert_eq!(metric.get(), 1);
-    }
-
-    #[tokio::test]
-    #[tracing_test::traced_test]
-    async fn test_failed_build_validator_announces() {
-        let settings = generate_test_relayer_settings();
-
-        let registry = Registry::new();
-        let core_metrics = CoreMetrics::new("relayer", 4000, registry).unwrap();
-        let chain_metrics = ChainMetrics {
-            block_height: IntGaugeVec::new(
-                opts!("block_height", BLOCK_HEIGHT_HELP),
-                BLOCK_HEIGHT_LABELS,
-            )
-            .unwrap(),
-            gas_price: None,
-            critical_error: IntGaugeVec::new(
-                opts!("critical_error", CRITICAL_ERROR_HELP),
-                CRITICAL_ERROR_LABELS,
-            )
-            .unwrap(),
-        };
-
-        let mailboxes =
-            Relayer::build_validator_announces(&settings, &core_metrics, &chain_metrics).await;
-
-        assert_eq!(mailboxes.len(), 1);
-        assert!(mailboxes.contains_key(&HyperlaneDomain::Known(KnownHyperlaneDomain::Arbitrum)));
-
-        // Arbitrum chain should not have any errors because it's ChainConf exists
-        let metric = chain_metrics
-            .critical_error
-            .get_metric_with_label_values(&["arbitrum"])
-            .unwrap();
-        assert_eq!(metric.get(), 0);
-
-        // Ethereum chain should error because it is missing ChainConf
-        let metric = chain_metrics
-            .critical_error
-            .get_metric_with_label_values(&["ethereum"])
-            .unwrap();
-        assert_eq!(metric.get(), 1);
-
-        // Optimism chain should error because it is missing ChainConf
-        let metric = chain_metrics
-            .critical_error
-            .get_metric_with_label_values(&["optimism"])
-            .unwrap();
-        assert_eq!(metric.get(), 1);
-    }
-}
+mod tests;