+++ rust/main/chains/hyperlane-starknet/src/mailbox.rs
@@ -1,8 +1,6 @@
 #![allow(clippy::enum_variant_names)]
 #![allow(missing_docs)]
 
-use std::sync::Arc;
-
 use byteorder::{BigEndian, ByteOrder};
 
 use async_trait::async_trait;
@@ -12,11 +10,10 @@ use hyperlane_core::{
     HyperlaneDomain, HyperlaneMessage, HyperlaneProvider, Mailbox, TxCostEstimate, TxOutcome, H256,
     U256,
 };
-use hyperlane_core::{FixedPointNumber, ReorgPeriod};
-use starknet::accounts::{ExecutionV3, SingleOwnerAccount};
+use hyperlane_core::{BatchItem, BatchResult, FixedPointNumber, QueueOperation, ReorgPeriod};
+use starknet::accounts::{Account, ExecutionV3, SingleOwnerAccount};
 use starknet::core::types::Felt;
 
-use starknet::providers::AnyProvider;
 use starknet::signers::LocalWallet;
 use tracing::instrument;
 
@@ -25,14 +22,14 @@ use crate::error::HyperlaneStarknetError;
 use crate::types::HyH256;
 use crate::{
     build_single_owner_account, get_block_height_for_reorg_period, send_and_confirm,
-    ConnectionConf, Signer, StarknetProvider,
+    ConnectionConf, JsonProvider, Signer, StarknetProvider,
 };
 
 /// A reference to a Mailbox contract on some Starknet chain
 #[derive(Debug)]
 #[allow(unused)]
 pub struct StarknetMailbox {
-    contract: Arc<StarknetMailboxInternal<SingleOwnerAccount<AnyProvider, LocalWallet>>>,
+    contract: StarknetMailboxInternal<SingleOwnerAccount<JsonProvider, LocalWallet>>,
     provider: StarknetProvider,
     conn: ConnectionConf,
 }
@@ -45,14 +42,14 @@ impl StarknetMailbox {
         locator: &ContractLocator<'_>,
         signer: Option<Signer>,
     ) -> ChainResult<Self> {
-        let account = build_single_owner_account(&conn.url, signer).await?;
+        let account = build_single_owner_account(conn.urls.clone(), signer).await?;
 
         let mailbox_address: Felt = HyH256(locator.address).into();
 
         let contract = StarknetMailboxInternal::new(mailbox_address, account);
 
         Ok(Self {
-            contract: Arc::new(contract),
+            contract,
             provider: StarknetProvider::new(locator.domain.clone(), conn),
             conn: conn.clone(),
         })
@@ -64,15 +61,8 @@ impl StarknetMailbox {
         &self,
         message: &HyperlaneMessage,
         metadata: &[u8],
-    ) -> ChainResult<ExecutionV3<'_, SingleOwnerAccount<AnyProvider, LocalWallet>>> {
+    ) -> ChainResult<ExecutionV3<'_, SingleOwnerAccount<JsonProvider, LocalWallet>>> {
         Ok(self.contract.process(&metadata.into(), &message.into()))
-    }
-
-    #[allow(unused)]
-    pub fn contract(
-        &self,
-    ) -> &StarknetMailboxInternal<SingleOwnerAccount<AnyProvider, LocalWallet>> {
-        &self.contract
     }
 }
 
@@ -97,7 +87,7 @@ impl Mailbox for StarknetMailbox {
     #[instrument(skip(self))]
     async fn count(&self, reorg_period: &ReorgPeriod) -> ChainResult<u32> {
         let block_number =
-            get_block_height_for_reorg_period(&self.provider.rpc_client(), reorg_period).await?;
+            get_block_height_for_reorg_period(self.provider.rpc_client(), reorg_period).await?;
 
         let nonce = self
             .contract
@@ -154,7 +144,7 @@ impl Mailbox for StarknetMailbox {
         _tx_gas_limit: Option<U256>,
     ) -> ChainResult<TxOutcome> {
         let contract_call = self.process_contract_call(message, metadata).await?;
-        send_and_confirm(&self.provider.rpc_client(), contract_call).await
+        send_and_confirm(self.provider.rpc_client(), contract_call).await
     }
 
     #[instrument(skip(self), fields(msg=%message, metadata=%bytes_to_hex(metadata)))]
@@ -190,5 +180,39 @@ impl Mailbox for StarknetMailbox {
 
     fn delivered_calldata(&self, _message_id: H256) -> ChainResult<Option<Vec<u8>>> {
         todo!()
+    }
+
+    /// True if the destination chain supports batching
+    /// (i.e. if the mailbox contract will succeed on a `process_batch` call)
+    fn supports_batching(&self) -> bool {
+        true
+    }
+
+    /// Try process the given operations as a batch. Returns the outcome of the
+    /// batch (if one was submitted) and the operations that were not submitted.
+    async fn process_batch<'a>(&self, ops: Vec<&'a QueueOperation>) -> ChainResult<BatchResult> {
+        let messages = ops
+            .iter()
+            .map(|op| op.try_batch())
+            .collect::<ChainResult<Vec<BatchItem<HyperlaneMessage>>>>()?;
+
+        let calls: Vec<_> = messages
+            .iter()
+            .map(|item| {
+                let metadata = item.submission_data.metadata.as_slice();
+                let message = &item.data;
+                self.contract
+                    .process_getcall(&metadata.into(), &message.into())
+            })
+            .collect();
+
+        let tx = self.contract.account.execute_v3(calls);
+        let outcome = send_and_confirm(self.provider.rpc_client(), tx).await?;
+
+        // Either all operations are executed successfully, or none of them are
+        Ok(BatchResult {
+            outcome: Some(outcome),
+            failed_indexes: vec![],
+        })
     }
 }