+++ rust/main/utils/run-locally/src/ethereum/mod.rs
@@ -12,8 +12,8 @@ use crate::logging::log;
 use crate::program::Program;
 use crate::utils::{as_task, get_ts_infra_path, get_workspace_path, AgentHandles, TaskHandle};
 
+pub mod ethereum_termination_invariants;
 mod multicall;
-pub mod termination_invariants;
 
 #[apply(as_task)]
 pub fn start_anvil(config: Arc<Config>) -> AgentHandles {